refactor: auth middlewares in separate files
